// src/models/commands.enum.ts
enum ECOMMANDS {
    PN532_COMMAND_DIAGNOSE = 0x00,
    PN532_COMMAND_GETFIRMWAREVERSION = 0x02,
    PN532_COMMAND_GETGENERALSTATUS = 0x04,
    PN532_COMMAND_READREGISTER = 0x06,
    PN532_COMMAND_WRITEREGISTER = 0x08,
    PN532_COMMAND_READGPIO = 0x0C,
    PN532_COMMAND_WRITEGPIO = 0x0E,
    PN532_COMMAND_SETSERIALBAUDRATE = 0x10,
    PN532_COMMAND_SETPARAMETERS = 0x12,
    PN532_COMMAND_SAMCONFIGURATION = 0x14,
    PN532_COMMAND_POWERDOWN = 0x16,
    PN532_COMMAND_RFCONFIGURATION = 0x32,
    PN532_COMMAND_RFREGULATIONTEST = 0x58,
    PN532_COMMAND_INJUMPFORDEP = 0x56,
    PN532_COMMAND_INJUMPFORPSL = 0x46,
    PN532_COMMAND_INLISTPASSIVETARGET = 0x4A,
    PN532_COMMAND_INATR = 0x50,
    PN532_COMMAND_INPSL = 0x4E,
    PN532_COMMAND_INDATAEXCHANGE = 0x40,
    PN532_COMMAND_INCOMMUNICATETHRU = 0x42,
    PN532_COMMAND_INDESELECT = 0x44,
    PN532_COMMAND_INRELEASE = 0x52,
    PN532_COMMAND_INSELECT = 0x54,
    PN532_COMMAND_INAUTOPOLL = 0x60,
    PN532_COMMAND_TGINITASTARGET = 0x8C,
    PN532_COMMAND_TGSETGENERALBYTES = 0x92,
    PN532_COMMAND_TGGETDATA = 0x86,
    PN532_COMMAND_TGSETDATA = 0x8E,
    PN532_COMMAND_TGSETMETADATA = 0x94,
    PN532_COMMAND_TGGETINITIATORCOMMAND = 0x88,
    PN532_COMMAND_TGRESPONSETOINITIATOR = 0x90,
    PN532_COMMAND_TGGETTARGETSTATUS = 0x8A,

    PN532_RESPONSE_INDATAEXCHANGE = 0x41,
    PN532_RESPONSE_INLISTPASSIVETARGET = 0x4B,


    PN532_MIFARE_ISO14443A = 0x00,

    // Mifare Commands
    MIFARE_CMD_AUTH_A = 0x60,
    MIFARE_CMD_AUTH_B = 0x61,
    MIFARE_CMD_READ = 0x30,
    MIFARE_CMD_WRITE = 0xA0,
    MIFARE_CMD_WRITE_ULTRALIGHT = 0xA2,
    MIFARE_CMD_TRANSFER = 0xB0,
    MIFARE_CMD_DECREMENT = 0xC0,
    MIFARE_CMD_INCREMENT = 0xC1,
    MIFARE_CMD_STORE = 0xC2,

    // FeliCa Commands
    FELICA_CMD_POLLING = 0x00,
    FELICA_CMD_REQUEST_SERVICE = 0x02,
    FELICA_CMD_REQUEST_RESPONSE = 0x04,
    FELICA_CMD_READ_WITHOUT_ENCRYPTION = 0x06,
    FELICA_CMD_WRITE_WITHOUT_ENCRYPTION = 0x08,
    FELICA_CMD_REQUEST_SYSTEM_CODE = 0x0C,

    // Prefixes for NDEF Records =to identify record type,
    NDEF_URIPREFIX_NONE = 0x00,
    NDEF_URIPREFIX_HTTP_WWWDOT = 0x01,
    NDEF_URIPREFIX_HTTPS_WWWDOT = 0x02,
    NDEF_URIPREFIX_HTTP = 0x03,
    NDEF_URIPREFIX_HTTPS = 0x04,
    NDEF_URIPREFIX_TEL = 0x05,
    NDEF_URIPREFIX_MAILTO = 0x06,
    NDEF_URIPREFIX_FTP_ANONAT = 0x07,
    NDEF_URIPREFIX_FTP_FTPDOT = 0x08,
    NDEF_URIPREFIX_FTPS = 0x09,
    NDEF_URIPREFIX_SFTP = 0x0A,
    NDEF_URIPREFIX_SMB = 0x0B,
    NDEF_URIPREFIX_NFS = 0x0C,
    NDEF_URIPREFIX_FTP = 0x0D,
    NDEF_URIPREFIX_DAV = 0x0E,
    NDEF_URIPREFIX_NEWS = 0x0F,
    NDEF_URIPREFIX_TELNET = 0x10,
    NDEF_URIPREFIX_IMAP = 0x11,
    NDEF_URIPREFIX_RTSP = 0x12,
    NDEF_URIPREFIX_URN = 0x13,
    NDEF_URIPREFIX_POP = 0x14,
    NDEF_URIPREFIX_SIP = 0x15,
    NDEF_URIPREFIX_SIPS = 0x16,
    NDEF_URIPREFIX_TFTP = 0x17,
    NDEF_URIPREFIX_BTSPP = 0x18,
    NDEF_URIPREFIX_BTL2CAP = 0x19,
    NDEF_URIPREFIX_BTGOEP = 0x1A,
    NDEF_URIPREFIX_TCPOBEX = 0x1B,
    NDEF_URIPREFIX_IRDAOBEX = 0x1C,
    NDEF_URIPREFIX_FILE = 0x1D,
    NDEF_URIPREFIX_URN_EPC_ID = 0x1E,
    NDEF_URIPREFIX_URN_EPC_TAG = 0x1F,
    NDEF_URIPREFIX_URN_EPC_PAT = 0x20,
    NDEF_URIPREFIX_URN_EPC_RAW = 0x21,
    NDEF_URIPREFIX_URN_EPC = 0x22,
    NDEF_URIPREFIX_URN_NFC = 0x23,

    PN532_GPIO_VALIDATIONBIT = 0x80,
    PN532_GPIO_P30 = 0,
    PN532_GPIO_P31 = 1,
    PN532_GPIO_P32 = 2,
    PN532_GPIO_P33 = 3,
    PN532_GPIO_P34 = 4,
    PN532_GPIO_P35 = 5,

    // FeliCa consts
    FELICA_READ_MAX_SERVICE_NUM = 16,
    FELICA_READ_MAX_BLOCK_NUM = 12, // for typical FeliCa card
    FELICA_WRITE_MAX_SERVICE_NUM = 16,
    FELICA_WRITE_MAX_BLOCK_NUM = 10, // for typical FeliCa card
    FELICA_REQ_SERVICE_MAX_NODE_NUM = 32,
    PREAMBLE = 0x00,
    START_CODE_1 = 0x00,
    START_CODE_2 = 0xff,
    POSTAMBLE = 0x00,

    SAMCONFIGURATION_MODE_NORMAL = 0x01,
    SAMCONFIGURATION_MODE_VIRTUAL_CARD = 0x02,
    SAMCONFIGURATION_MODE_WIRED_CARD = 0x03,
    SAMCONFIGURATION_MODE_DUAL_CARD = 0X04,
}